GUIScripts: setup stdio redirection in Main.py
we should do this ASAP, also, it wasn't working for me so I adjusted the wrapper to implement flush()
